# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Steak API Documentation
  description: Steak is a full-stack platform inspired by Steam and Epic Games, designed for buying,
    downloading, and managing licensed PC games. It supports both players and game publishers.
  version: 1.0.0
servers:
  - url: 'https://api.steak.io.vn/'
  - url: 'http://localhost:8888/'
...
paths:
  /api/v1/user/auth/register:
    post:
      tags:
        - user
      summary: Gửi yêu cầu tạo tài khoản người dùng
      description: Gửi yêu cầu tạo tài khoản người dùng (kiểm tra thông tin và gửi email xác nhận)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml#/components/schemas/UserRegisterRequest"
      responses:
        200:
          description: Thông tin tài khoản hợp lệ, check mail để xác nhận tài khoản
        409:
          description: Thông tin tài khoản đã tồn tại trên hệ thống
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
  /api/v1/user/auth/username-login:
    post:
      tags:
        - user
      summary: Đăng nhập tài khoản
      description: Đăng nhập vào tài khoản bằng username
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/UsernameLoginRequest"

      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/UserLoginResponse"
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/user/auth/email-login:
    post:
      tags:
        - user
      summary: Đăng nhập tài khoản
      description: Đăng nhập vào tài khoản bằng email
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/EmailLoginRequest"

      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/UserLoginResponse"
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/auth/register:
    post:
      tags:
        - dev
      summary: Đăng ký nhà phát hành
      description: Đăng ký nhà phát hành
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/PublisherRegisterRequest"
      responses:
        200:
          description: Thông tin tài khoản hợp lệ, check mail để xác nhận tài khoản
        409:
          description: Thông tin tài khoản đã tồn tại trên hệ thống
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
  /api/v1/dev/auth/username-login:
    post:
      tags:
        - dev
      summary: Đăng nhập tài khoản nhà phát hành
      description: Đăng nhập vào tài khoản
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/UsernameLoginRequest"
      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: publisher123
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/auth/email-login:
    post:
      tags:
        - dev
      summary: Đăng nhập tài khoản nhà phát hành
      description: Đăng nhập vào tài khoản bằng email
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/EmailLoginRequest"
      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: publisher123
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/admin/auth/username-login:
    post:
      tags:
        - admin
      summary: Đăng nhập vào tài khoản quản trị
      description: Đăng nhập vào tải khoản quản trị bằng username
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/UsernameLoginRequest"
      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: admin123
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/admin/auth/email-login:
    post:
      tags:
        - admin
      summary: Đăng nhập vào tài khoản quản trị
      description: Đăng nhập vào tải khoản quản trị bằng email
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/EmailLoginRequest"
      responses:
        200:
          description: Đăng nhập thành công, token và refreshToken được lưu vào cookie set full cờ
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: admin123@gmail.com
        default:
          description: Đăng nhập thất bại
          content:
            application/json:
              schema:
                $ref: "./components.yaml#/components/schemas/Error"
  /api/v1/dev/upload/presigned-image-url:
    post:
      tags:
        - publisher-upload
      summary: Lấy URL để upload ảnh
      description: Lấy URL để upload ảnh cho game
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/ImageUploadPresignRequest"
      responses:
        200:
          description: Trả về URL để upload ảnh
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/ImageUploadPresignResponse"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/upload/presigned-image-urls:
    post:
      tags:
        - publisher-upload
      summary: Lấy URL để upload nhiều ảnh
      description: Lấy URL để upload nhiều ảnh cho game
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "./components.yaml/#/components/schemas/ImageUploadPresignRequest"
      responses:
        200:
          description: Trả về URL để upload ảnh
          content:
            application/json:
              schema:
                type: object
                properties:
                  urls:
                    type: array
                    items:
                      $ref: "./components.yaml/#/components/schemas/ImageUploadPresignResponse"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/upload/delete-image:
    delete:
      tags:
        - publisher-upload
      summary: Xoá ảnh đã upload
      description: Xoá ảnh đã upload trên S3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/ImageDeleteRequest"
      responses:
        200:
          description: Xoá ảnh thành công
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/upload/delete-images:
    delete:
      tags:
        - publisher-upload
      summary: Xoá nhiều ảnh đã upload
      description: Xoá nhiều ảnh đã upload trên S3
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "./components.yaml/#/components/schemas/ImageDeleteRequest"
      responses:
        200:
          description: Xoá ảnh thành công
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/upload/presigned-part-game-url:
    post:
      tags:
        - publisher-upload
      summary: Lấy URL upload từng phần cho file game lớn
      description: Lấy presigned URL để upload từng phần của file game (multipart upload)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/GameUploadPresignedRequest"
      responses:
        200:
          description: Trả về thông tin presigned URL cho từng part
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/PresignedUrlResponse"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/upload/recreate-presigned-upload-url:
    post:
      tags:
        - publisher-upload
      summary: Tạo lại presigned URL upload cho file game
      description: Tạo lại presigned URL upload cho file game khi URL cũ hết hạn
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/RecreatePresignedUrlRequest"
      responses:
        200:
          description: Trả về presigned URL mới
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/PresignedUrlResponse"
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
  /api/v1/dev/upload/complete-part-upload:
    post:
      tags:
        - publisher-upload
      summary: Hoàn thành upload multipart file game
      description: Thông báo đã upload xong các part, hoàn thành upload file game lớn
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./components.yaml/#/components/schemas/CompleteMultipartRequest"
      responses:
        200:
          description: Hoàn thành upload thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload completed successfully
        default:
          description: Lỗi khác
          content:
            application/json:
              schema:
                $ref: "./components.yaml/#/components/schemas/Error"
...